feat: Stream while downloading when supported

+++ src/server/http-handler/video.ts
@@ -1,39 +1,62 @@
 import fs from 'fs'
 import { animeVideoRealDirPath as realAnimeVideoDirPath } from '~s/utils/path'
 import { parseNumber } from '~/shared/utils/number'
+import { downloadSizeMap } from '~s/external/download/progress'
+import { withoutExtension } from '~/shared/utils/file'
 
 export const handleVideoRequest = async (request: Request, path: string): Promise<Response> => {
-  const pathArr = path.slice('videos/'.length).split('/')
+  const pathArr = path.slice('videos/'.length).split('/') as [string, string]
   if (pathArr.length !== 2) {
-    return new Response('Not found', { status: 404 })
+    return notFound()
   }
 
   const videoFileResult = await getVideoFile(pathArr as [string, string])
   if (!videoFileResult) {
-    return new Response('Not found', { status: 404 })
+    return notFound()
   }
 
-  const [video, videoPath] = videoFileResult
+  const [video, videoPath, isDownloading] = videoFileResult
   const range = request.headers.get('range')
   if (!range) {
     return new Response(video)
   }
 
+  let size = isDownloading ? downloadSizeMap.get(withoutExtension(pathArr.join(':'))) : video.size
+  if (!size) {
+    return notFound()
+  }
+
   const parts = range.replace('bytes=', '').split('-')
   const start = parseNumber(parts[0]) ?? 0
-  const end = parseNumber(parts[1]) ?? video.size - 1
+  const end = Math.min(parseNumber(parts[1]) ?? Infinity, video.size - 1)
 
-  const res = new Response(fs.createReadStream(videoPath, { start, end }), { status: 206 })
+  try {
+    const res = new Response(fs.createReadStream(videoPath, { start, end }), { status: 206 })
 
-  res.headers.set('Content-Range', `bytes ${start}-${end}/${video.size}`)
-  res.headers.set('Accept-Ranges', 'bytes')
-  res.headers.set('Content-Length', String(end - start + 1))
-  res.headers.set('Content-Type', video.type)
+    res.headers.set('Content-Range', `bytes ${start}-${end}/${size}`)
+    res.headers.set('Accept-Ranges', 'bytes')
+    res.headers.set('Content-Length', String(end - start + 1))
+    res.headers.set('Content-Type', video.type)
 
-  return res
+    return res
+  } catch {
+    const res = new Response(null, { status: 416 })
+
+    res.headers.set('Content-Range', `bytes */${video.size}`)
+    res.headers.set('Content-Type', video.type)
+
+    return res
+  }
+
+  function notFound() {
+    return new Response('Not found', { status: 404 })
+  }
 }
 
-const getVideoFile = async ([animeId, episodeNumber]: [string, string], _retry = true) => {
+const getVideoFile = async (
+  [animeId, episodeNumber]: [string, string],
+  _retry = true,
+): Promise<[Bun.BunFile, string, boolean] | null> => {
   const [realVideoDirPath, videoPathIsFromCache] = await getRealVideoDirPath(animeId)
   if (!realVideoDirPath) {
     return null
@@ -43,6 +66,12 @@ const getVideoFile = async ([animeId, episodeNumber]: [string, string], _retry =
   const video = Bun.file(videoPath)
 
   if (video.size === 0) {
+    const downloadingVideoPath = videoPath.replace('.mp4', '_.mp4')
+    const downloadingVideo = Bun.file(downloadingVideoPath)
+    if (downloadingVideo.size) {
+      return [downloadingVideo, downloadingVideoPath, true]
+    }
+
     if (videoPathIsFromCache && _retry) {
       realVideoDirPathCache.delete(animeId)
 
@@ -52,14 +81,14 @@ const getVideoFile = async ([animeId, episodeNumber]: [string, string], _retry =
     return null
   }
 
-  return [video, videoPath] as const
+  return [video, videoPath, false]
 }
 
 const realVideoDirPathCache = new Map<string, string>()
 const getRealVideoDirPath = async (animeId: string) => {
   const cache = realVideoDirPathCache.get(animeId)
   if (cache) {
-    return [cache, true]
+    return [cache, true] as const
   }
 
   const result = await realAnimeVideoDirPath(animeId)
@@ -67,5 +96,5 @@ const getRealVideoDirPath = async (animeId: string) => {
     realVideoDirPathCache.set(animeId, result)
   }
 
-  return [result, false]
+  return [result, false] as const
 }
